+++ frontend/src/components/AlpineWrapper.astro
@@ -1,5 +1,7 @@
 <script>
-  window.API_ENDPOINT = 'http://localhost:5000/app';
+  window.API_ENDPOINT = 'http://localhost:5000/api';
+  window.AUTH_ENDPOINT = 'http://localhost:5000/auth';
+  window.BASE_URL = "http://localhost:4321/proyecto_agua/";
   
   import Alpine from 'alpinejs';
